requirements.txt
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
asyncpg==0.30.0
blinker==1.9.0
boto3==1.36.18
botocore==1.36.18
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-mail==1.4.2
fastapi-pagination==0.12.34
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
loguru==0.7.3
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.2
orjson==3.10.15
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.2
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
s3transfer==0.11.2
secure-smtplib==0.1.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-id==1.0.2
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
